Refactor settings component***
