--- app/Http/Livewire/SettingsComponent.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Http\Livewire;
-
-use Livewire\Component;
-
-class SettingsComponent extends Component
-{
-    public function render()
-    {
-        return view('livewire.settings-component');
-    }
-}
