--- resources/views/livewire/settings-component.blade.php
@@ -1,5 +1,0 @@
-@role('app.admin')
-    @livewire('settings.app-settings-component')
-@else
-    @livewire('settings.tenant-settings-component')
-@endrole
